refactor: connection middleware refactoring
