+++ backend/src/middlewares/connectionMiddleware.ts
@@ -1,23 +1,11 @@
 import { NextFunction, Request, Response } from 'express';
-import { connect } from 'mongoose';
 
 import ServiceUnavailableError from 'classes/errors/serverErrors/ServiceUnavailableError';
+import { connectToDatabase } from 'utils/connection';
 
 const connectionMiddleware = async (req: Request, res: Response, next: NextFunction) => {
   try {
-    const DB_URL: Record<string, string | undefined> = {
-      local: process.env.DATABASE_LOCAL_URL,
-      dev: process.env.DATABASE_URL,
-      backdev: process.env.DATABASE_BACKDEV_URL,
-      test: process.env.DATABASE_URL,
-      production: process.env.DATABASE_URL,
-    };
-
-    const CONNECTION_STRING = process.env.NODE_ENV && DB_URL[`${process.env.NODE_ENV}`];
-    if (CONNECTION_STRING) {
-      await connect(CONNECTION_STRING);
-    }
-
+    await connectToDatabase();
     next();
   } catch (err) {
     const { message } = err as Error;
