+++ backend/src/utils/connection/index.ts
@@ -1,0 +1,1 @@
+export { default as connectToDatabase } from './connectToDatabase';
